+++ src/app/types/market.ts
@@ -1,0 +1,7 @@
+export type MarketSummary = {
+  market: string;
+  key: string;
+  price: string;
+  marketSize: string;
+  fee: string;
+};
